refactor: refactor and update styles
